Bus

Description

This module defines methods for the shared bus between processors

Documentation

newtype EventBus a

Constructors

EventBus

Fields

getSeq :: Seq a

data CacheEventBus

Constructors

CacheEventBus

Fields

getEventBus :: EventBus BusEvent

getCaches :: [Cache]

getMemory :: Memory

isFree :: Bool

isBusy :: Bool

data BusEvent

Constructors

MESIBusRd

MESIBusRdX

DragonBusRd

DragonBusUpd

createNewCacheEventBus :: [Cache] -> Memory -> CacheEventBus

Creating brand new cache event bus - likely to be only called once then updated with recreate method

recreateCacheEventBus :: CacheEventBus -> [Cache] -> CacheEventBus

Method to re-create the cache event bus on every new cycle

enqueueEventBus :: CacheEventBus -> BusEvent -> CacheEventBus

Public method for others to enqueue onto the cache event bus

acquireEventBus :: CacheEventBus -> Maybe CacheEventBus

Method for processors to attempt unique access to the bus Maybe return a new cache event bus to indicate we managed to lock the bus down for us

releaseEventBus :: CacheEventBus -> Maybe CacheEventBus

Method for a processor to give up control over the bus. ASSUMES WELL BEHAVED CALLS. No checking whether the processor that calls this method is authorized to do so! Maybe return a new cache event bus to indicate we managed to release the bus.

runEventBus :: CacheEventBus -> CacheEventBus

MAIN METHOD FOR BUS TO DEQUEUE AN EVENT AND EXECUTE IT TBI!

createBus :: EventBus a

enqueueBus :: EventBus a -> a -> EventBus a

dequeueBus :: EventBus a -> (EventBus a, a)

isBusEmpty :: EventBus a -> Bool