Cache

Documentation

data Cache

create :: CacheSize -> Associativity -> BlockSize -> Cache

Creates an empty cache with the specified cache size, associativity, and block size. Returns the empty cache on successful execution.

issueRead :: MemoryAddress -> Cache -> Cache

Issues a cache read of a memory address on the specified cache. Returns the renewed cache.

issueWrite :: MemoryAddress -> Cache -> Cache

Issues a cache write of a memory address to the specified cache. Returns the renewed cache.

busGetBlockState :: MemoryAddress -> Cache -> Maybe BlockState

busSetBlockState :: BlockState -> MemoryAddress -> Cache -> Cache

Sets the block state of the block the memory address resides in to the specified block state. Returns the renewed cache.

busAllocate :: BlockState -> MemoryAddress -> Cache -> (Maybe BlockState, Cache)

Allocates a memory address on the specified cache. Returns the evicted block state (if any) and the renewed cache.

busEvict :: MemoryAddress -> Cache -> (BlockState, Cache)

Evicts a memory address from the specified cache. Returns the evicted block state and the renewed cache.

elapse :: Cache -> Cache

isCacheHit :: Cache -> Maybe IsCacheHit