CacheBlock

Documentation

data BlockState

Constructors

M

E

S

I

C

SC

D

SD

data CacheBlock

create :: BlockSize -> CacheBlock

Creates an empty cache block with the specified block size. Returns the empty cache block.

allocate :: BlockState -> BlockTag -> Offset -> MemoryAddress -> CacheBlock -> CacheBlock

evict :: CacheBlock -> (BlockState, CacheBlock)

isValid :: CacheBlock -> Bool

Checks whether this cache block is valid (state other than I). Returns True if this cache block is valid, False otherwise.

hasTag :: BlockTag -> CacheBlock -> Bool

Checks whether this cache block has the specified tag. Returns True if this cache block has the specified tag, False otherwise.

getBlockState :: CacheBlock -> BlockState

setBlockState :: BlockState -> CacheBlock -> CacheBlock