CacheSet

Documentation

data CacheSet

create :: Associativity -> BlockSize -> CacheSet

Creates an empty cache set with the specified associativity and block size. Returns the empty cache set.

canAllocate :: CacheSet -> Bool

Checks whether a block in this cache is available for allocation. Returns True if there is an empty block, False otherwise.

allocate :: BlockState -> BlockTag -> Offset -> MemoryAddress -> CacheSet -> CacheSet

Allocates a cache block with given state/tag on the first available block in the specified cache set. Returns the renewed cache set.

evictLRU :: CacheSet -> (BlockState, CacheSet)

Evicts the first valid cache block within the specified cache set. Returns the renewed cache set.

evict :: BlockTag -> CacheSet -> (BlockState, CacheSet)

Evicts the cache block with the specified tag on the specified cache set. Returns the renewed cache set.

hasTag :: BlockTag -> CacheSet -> Bool

Finds a tag in a valid block within the specified cache set. Returns a boolean indicating whether the tag was found.

getBlockState :: BlockTag -> CacheSet -> Maybe BlockState

Retrieves the block state corresponding to the specified block tag on the specified cache set. Returns the block state if found, Nothing otherwise.

setBlockState :: BlockState -> BlockTag -> CacheSet -> CacheSet

Sets the state of the cache block with the given block tag to the given state. Returns the renewed cache set.