MemoryAddress

Documentation

parse :: CacheParams -> MemoryAddress -> (BlockTag, SetIndex, Offset)

Parses a memory address into the respective block tag, set index, and offset while adhering to the specified block size and number of cache sets. Returns the block tag, set index, and offset in a tuple.