SimulatorCore

Documentation

num_processors :: Int

runSimulation :: ProtocolInput -> Filename -> CacheSize -> Associativity -> BlockSize -> IO StatsReport

Given appropriate cmd line arguments, runs the entire cache simulation and results the stats results Reads in all traces and kicks off the pure part of this simulation

startSimulationPure :: [Processor] -> [[Trace]] -> StatsReport

Runs a single cycle for a single processor. Attempts to feed it a list of traces. The processor may consume some part of the list of traces, then return a new Processor instance (the new processor state essentially), the remaining traces to be consumed, and a list of bus events generated by this processor runOneProcessorCycle :: Processor -> [Trace] -> (Processor, [Trace], [Message]) runOneProcessorCycle processor tracexs =

Start of the pure part of the code - the IO for the traces has been unwrapped in runSimulation

runAllSimulationCycles :: [(Processor, [Trace])] -> CacheEventBus -> Int -> Int -> SimulationStatistics

Pass in the processors and traces, tbe index of the current processor being worked on, and the number of cycles completed. |Eventually the statistics report will be returned

runOneProcessorCycle :: (Processor, [Trace]) -> CacheEventBus -> (Processor, [Trace], CacheEventBus)

Attempts to feed one trace to the processor, which it can avoid consuming if it's already working on something/busy Returns a new processor with updated state, the remaining traces to execute, and any new bus events to propagate

allProcessorsComplete :: [(Processor, [Trace])] -> Bool

getStatsReport :: [(Processor, [Trace])] -> Int -> SimulationStatistics