Trace

Documentation

toTraceWithError :: String -> Trace

This function converts a String to a valid trace. It can cause an error if it is called with an input that is not of the right format.

data Trace

Define the structure of the main source of input into the program: a single trace line. Can either be a load, store or other instructions where we wait for a certain number of cycles.

Constructors

LoadInstruction Address

StoreInstruction Address

OtherInstruction NumCycles